requirements.txt
torch==2.1.2 --index-url https://download.pytorch.org/whl/cpu
transformers
sentence-transformers
numpy
pandas
ipykernel
matplotlib
seaborn
pymysql
python-dotenv
nltk
emoji
textblob
langchain
langchain_community
openai
langchain-openai
langchain-chroma
pypdf
streamlit
langchain_huggingface
groq
langchain_groq